+++ packages/remix/src/server/index.ts
@@ -145,3 +145,4 @@ export { init, getRemixDefaultIntegrations } from './sdk';
 export { captureRemixServerException } from './errors';
 export { sentryHandleError, wrapHandleErrorWithSentry, instrumentBuild } from './instrumentServer';
 export { generateSentryServerTimingHeader } from './serverTimingTracePropagation';
+export { remixIntegration } from './integrations/RemixIntegration';